Duelist Abilities supporting Elite Dexterity now use a dedicated class instead of using a ton of lambdas and a complicated method.
This should make it easier to understand and work with when future bugs come up.

* added an argument-less duelistAbility that returns the method

Redid inheritance of some weapons:
- Accurate Weapons are now subclasses of Mace
- Gloves and Gauntlet are now subclasses of Sai
- Swords are now subclasses of Sword
- Glaive is now a subclass of Spear

Fixed missile weapons proccing enchantments and doing damage as if they were +0 when equipped via elite dexterity. They now instead delegate to the weapon's proc method instead of skipping to the enchantment part.